requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-google-genai>=0.0.3
langgraph>=0.0.15
pydantic>=2.0.0
python-dotenv>=1.0.0
pyautogui>=0.9.54
scikit-learn>=1.2.2
langchain_community>=0.0.1
langchain_google_vertexai>=0.0.1
langchain_huggingface>=0.0.1
schedule>=1.2.0
mss>=6.2.0
plyer>=1.0.0
google.generativeai>=0.6.0

# Speech processing
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
playsound>=1.3.0
kokoro>=0.1.2
soundfile>=0.12.1
torch>=2.0.0

# UI frameworks
streamlit>=1.24.0
flask>=2.2.0

# API integrations
spotipy>=2.22.1
notion-client>=2.0.0
google-api-python-client>=2.70.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
groq>=0.4.0
numpy>=1.24.0
pillow>=9.5.0
opencv-python>=4.7.0
keyboard>=0.13.5
faiss-cpu>=1.7.4
tweepy>=4.12.1
jigsawstack>=0.0.1
filelock>=3.12.0
